offer workflow splitted into list and detail, administration workflow created
