+++ src/bot/Elrond.TradeOffer.Web/BotWorkflows/CommonQueries.cs
@@ -8,6 +8,7 @@ internal static class CommonQueries
     public const string PlaceBidQueryPrefix = "PlaceBid_";
     public const string ChangeNetworkOrAddressQuery = "changeNetworkOrAddress";
     public const string BackToHomeQuery = "BackToHome";
+    public const string AdministrationQuery = "administration";
 
     public static string ShowOfferQuery(Guid offerId) => $"{ShowOfferQueryPrefix}{offerId}";
 }
