+++ src/bot/Elrond.TradeOffer.Web/BotWorkflows/Workflows/IOfferListNavigation.cs
@@ -3,9 +3,7 @@ using Telegram.Bot;
 
 namespace Elrond.TradeOffer.Web.BotWorkflows.Workflows;
 
-public interface IOfferNavigation
+public interface IOfferListNavigation
 {
     Task ShowOffersAsync(ITelegramBotClient client, long userId, long chatId, OfferFilter filter, CancellationToken ct);
-
-    Task ShowOfferAsync(ITelegramBotClient client, long userId, long chatId, Guid offerId, CancellationToken ct);
 }
